requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
flask-wtf
flask-mail
PyMySQL==0.6.2
Werkzeug==0.9.4
